+++ farmacia-servicio/src/main/java/com/f_rafael/farmacia_servicio/service/DosisService.java
@@ -1,5 +1,7 @@
 package com.f_rafael.farmacia_servicio.service;
 
+import com.f_rafael.farmacia_servicio.exception.CampoNuloException;
+import com.f_rafael.farmacia_servicio.exception.EntidadNoEncontradaException;
 import com.f_rafael.farmacia_servicio.model.Dosis;
 import com.f_rafael.farmacia_servicio.repository.IDosisRepository;
 import lombok.AllArgsConstructor;
@@ -20,6 +22,16 @@ public class DosisService implements IDosisService{
     }
 
     @Override
+    public Dosis buscarPorId2(Long id) {
+
+        if(repository.findById(id).isEmpty()){
+            throw new EntidadNoEncontradaException("Entidad no encontrada");
+        }
+
+        return repository.findById(id).get();
+    }
+
+    @Override
     public List<Dosis> buscarTodas() {
         return repository.findAll();
     }
@@ -35,12 +47,38 @@ public class DosisService implements IDosisService{
     }
 
     @Override
+    public Dosis actualizar2(Dosis dosis) {
+        Long id = dosis.getId();
+
+        if(id == null || dosis.getCantidad() == null || dosis.getUnidad() == null || dosis.getIntervaloHoras() == null){
+            throw new CampoNuloException("Ningún campo puede ser nulo");
+        }
+
+        return this.guardar(dosis);
+    }
+
+    @Override
     public void borrarPorId(Long id) {
+
+        if(repository.findById(id).isEmpty()){
+            throw new EntidadNoEncontradaException("Entidad no encontrada");
+        }
+
         repository.deleteById(id);
     }
 
     @Override
     public Optional<Dosis> buscarPorCantidadUnidadEIntervalo(float cantidad, String nombreUnidad, int intervalo) {
         return repository.buscarPorCantidadUnidadEIntervalo(cantidad,nombreUnidad,intervalo);
+    }
+
+    @Override
+    public Dosis buscarPorCantidadUnidadEIntervalo2(float cantidad, String nombreUnidad, int intervalo) {
+
+        if(repository.buscarPorCantidadUnidadEIntervalo(cantidad,nombreUnidad,intervalo).isEmpty()){
+            throw new EntidadNoEncontradaException("Entidad no encontrada");
+        }
+
+        return repository.buscarPorCantidadUnidadEIntervalo(cantidad,nombreUnidad,intervalo).get();
     }
 }
